Fixing up Map collission code. Map collission comes properly from the data.

+++ src/ADV_OnCleanup.cpp
@@ -12,7 +12,7 @@ void AdvGame::OnCleanup() {
     //fprintf (pFile, "Reached Point 1\n");
     //fclose (pFile); 
     
-	for(int i = 0; i < CEntity::EntityList.size(); i++) {
+	for(Uint16 i = 0; i < CEntity::EntityList.size(); i++) {
             if(!CEntity::EntityList[i]) continue;
             
             CEntity::EntityList[i]->OnCleanup();
